+++ compatibility-tests/compile-fail/tests/ui/duplication-source-attributes.rs
@@ -1,15 +1,15 @@
- use snafu::Snafu;
+use snafu::Snafu;
 
- #[derive(Debug, Snafu)]
- enum EnumError {
-     AVariant {
-         // First source, legitimate
-         source: String,
+#[derive(Debug, Snafu)]
+enum EnumError {
+    AVariant {
+        // First source, legitimate
+        source: String,
 
-         // Should mark this second source as a duplicate
-         #[snafu(source)]
-         my_source: String,
-     },
- }
+        // Should mark this second source as a duplicate
+        #[snafu(source)]
+        my_source: String,
+    },
+}
 
 fn main() {}