+++ compatibility-tests/compile-fail/tests/ui/duplication-source-from-attributes.rs
@@ -1,13 +1,13 @@
- use snafu::Snafu;
+use snafu::Snafu;
 
- #[derive(Debug, Snafu)]
- enum EnumError {
-     AVariant {
-         // Should detect second attribute as duplicate
-         #[snafu(source(from(EnumError, Box::new)))]
-         #[snafu(source(from(EnumError, Box::new)))]
-         my_source: Box<EnumError>,
-     },
- }
+#[derive(Debug, Snafu)]
+enum EnumError {
+    AVariant {
+        // Should detect second attribute as duplicate
+        #[snafu(source(from(EnumError, Box::new)))]
+        #[snafu(source(from(EnumError, Box::new)))]
+        my_source: Box<EnumError>,
+    },
+}
 
 fn main() {}